src/processors/powerpoint_processor/requirements.txt
python-pptx==1.0.2